+++ scripts/eval_hr_tau_demo_eb_golden_v1.py
@@ -22,7 +22,7 @@ import argparse
 
 import pandas as pd
 
-from eb_examples.paths import GoldenV1Artifacts, resolve_base_dir
+from eb_examples import GoldenV1Artifacts, resolve_base_dir
 from eb_metrics import hr_at_tau